pyproject.toml
[project]
name = "context-shield"
version = "0.1.0"
description = "Middleware layer that enforces trust boundaries for autonomous AI agents"
requires-python = ">=3.10"
license = "MIT"
readme = "README.md"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
demo = ["anthropic>=0.40"]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
